decouple AbstractValueObjectRemoverRector
